Добавляет проверку на id элемента клика миниатюр
