+++ js/validate-form.js
@@ -1,7 +1,7 @@
-import {closeForm} from './close-form.js';
+import {closeForm} from './open-form.js';
 import {checkLengthComment, checkHashtags, checkLengthHashtags, checkRepeatHashtags} from './functions-validate-form.js';
 import {onEffectChengePicture} from './slider.js';
-import {onScalingAdd} from './scale-control.js';
+import {addScaling} from './scale-control.js';
 import {sendData} from './api.js';
 
 const imgEditorForm = document.querySelector('.img-upload__form');
@@ -73,7 +73,7 @@ const setUserFormSubmit = (onSuccess) => {
   });
 };
 
-onScalingAdd();
+addScaling();
 
 effectsList.addEventListener('change', onEffectChengePicture);
